Add table test
